% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_model.R
\name{analyse_model}
\alias{analyse_model}
\title{Analyse fitted models from \code{fit_model} and \code{fit_multiple_models}}
\usage{
analyse_model(
  model,
  pairwise.test = TRUE,
  Q.norm.factor = NA,
  interval_ahi = c(0, 24),
  DIC_treshold = 1,
  get.chains = FALSE
)
}
\arguments{
\item{model}{the output model from \code{fit_model} or \code{fit_multiple_models}}

\item{pairwise.test}{if test the pairwise differences between each group}

\item{Q.norm.factor}{normalization factor for calculating effect magnitide between treatment, if NA is calculated automatically ad 99 percentile of all Qs. Default: NA.}

\item{interval_ahi}{vector of 2 indicating the time interval to calculate AhI. Default: c(0,24)}

\item{DIC_treshold}{minimum difference needed in DIC to select the more complex model (Weibull). Default: 1.}

\item{get.chains}{get as output the parameter chains extracted from model. Defauls: FALSE.}
}
\description{
Analyse fitted models from \code{fit_model} and \code{fit_multiple_models}
}
\examples{
\dontrun{
data = data.frame(group = c("A","A","A","B","B","B"),
       time = c(1,2,24,1,2,24),
       moving = c(0,1,4,2,5,16),
       tot = rep(20,6))
mod1 = fit_model(data, model = "weibull")
analyse_model(mod1)

mod2 = fit_multiple_models(data)
analyse_model(mod2)}
}
