% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{fit_model_exponential}
\alias{fit_model_exponential}
\title{Fit Exponential model to anhydrobiosis recovery data}
\usage{
fit_model_exponential(
  data,
  column_IDs = c(1, 2, 3, 4),
  prop = 0.9,
  n.iter = 1e+06,
  ...
)
}
\arguments{
\item{data}{a data frame table with observations in rows and variables in columns}

\item{column_IDs}{a vector with the numeric indexes of the column containing in order: population, time, moving individuals, total individuals. Default: c(1,2,3,4).}

\item{prop}{numeric value between 0 and 1 indicating the quantile for the calculation of Q. Default: 0.90.}

\item{n.iter}{number of iterations for the JAGS model. Default: 1000000.}

\item{...}{additional arguments for \code{R2Jags::jags}.}
}
\value{
a JAGS model
}
\description{
Fit Exponential model to anhydrobiosis recovery data
}
