% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_multiple_models.R
\name{fit_multiple_models}
\alias{fit_multiple_models}
\title{Fit both Exponential and Weibull models to each experimental group}
\usage{
fit_multiple_models(
  data,
  column_IDs = c(1, 2, 3, 4),
  prop = 0.9,
  n.iter = 1e+06,
  ...
)
}
\arguments{
\item{data}{a data frame table with observations in rows and variables in columns}

\item{column_IDs}{a vector with the numeric indexes of the column containing in order: population, time, moving individuals, total individuals. Default: c(1,2,3,4).}

\item{prop}{numeric value between 0 and 1 indicating the quantile for the calculation of Q. Default: 0.90.}

\item{n.iter}{number of iterations for the JAGS model. Default: 1000000.}

\item{...}{additional arguments for \code{R2Jags::jags}.}
}
\value{
a list of models
}
\description{
Fit both Exponential and Weibull models to each experimental group
}
\examples{
\dontrun{
data = data.frame(group = c("A","A","A","B","B","B"),
       time = c(1,2,24,1,2,24),
       moving = c(0,1,4,2,5,16),
       tot = rep(20,6))
mod = fit_multiple_models(data)}
}
