% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_direction_2d.R
\name{p_direction_2d}
\alias{p_direction_2d}
\title{Calculated p_direction of differences in angular direction between two groups}
\usage{
p_direction_2d(group1, group2, make.plot = FALSE)
}
\arguments{
\item{group1}{a matrix with 2 columns with x and y coordinates}

\item{group2}{a matrix with 2 columns with x and y coordinates}

\item{make.plot}{make a direction plot of differences between groups. Defauls: FALSE.}
}
\value{
a p-direction value
}
\description{
Calculated p_direction of differences in angular direction between two groups
}
