% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fits.R
\name{plot_fits}
\alias{plot_fits}
\title{Make a plot showing predicted values along with the input data}
\usage{
plot_fits(model)
}
\arguments{
\item{model}{the output model from \code{fit_model} or \code{fit_multiple_models}}
}
\value{
a saved pdf file named model_fits.pdf
}
\description{
Make a plot showing predicted values along with the input data
}
