% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_model.R
\name{plot_model}
\alias{plot_model}
\title{Plot p and Q from fitted models from \code{fit_model} and \code{fit_multiple_models}}
\usage{
plot_model(model, interval_ahi = c(0, 24), DIC_treshold = 1)
}
\arguments{
\item{model}{the output model from \code{fit_model} or \code{fit_multiple_models}}

\item{interval_ahi}{vector of 2 indicating the time interval to calculate AhI. Default: c(0,24)}

\item{DIC_treshold}{minimum difference needed in DIC to select the more complex model (Weibull). Default: 1.}
}
\description{
Plot p and Q from fitted models from \code{fit_model} and \code{fit_multiple_models}
}
