% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fits.R
\name{plot_fits}
\alias{plot_fits}
\title{Save as pdf file in the working directory with plots showing the model fits}
\usage{
plot_fits(model)
}
\arguments{
\item{model}{the output model from \code{fit_model} or \code{fit_multiple_models}}
}
\value{
a saved pdf file mamed model_fits.pdf
}
\description{
Save as pdf file in the working directory with plots showing the model fits
}
